invoice-middleware/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
mysql-connector-python==9.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
alembic==1.13.3
tenacity==9.0.0
pymupdf==1.24.10
Pillow==10.4.0
openpyxl==3.1.5
reportlab==4.2.5
